std: Move platform specific memchr code into sys
