+++ src/libstd/sys/common/mod.rs
@@ -31,6 +31,7 @@ pub mod at_exit_imp;
 pub mod backtrace;
 pub mod condvar;
 pub mod io;
+pub mod memchr;
 pub mod mutex;
 pub mod net;
 pub mod poison;
